🚚 rename [#26] Comment 관련 파일의 경로 변경
