+++ src/features/card/components/card-detail-modal.tsx
@@ -12,9 +12,8 @@ import { ProfileSize } from "@/components/profile/profile-size";
 import Typography from "@/components/typography";
 import { deleteCard } from "@/features/card/apis";
 import { Direction, Menu, MenuItem } from "@/features/card/components/menu";
-import { getComments } from "@/features/comment/apis/comment";
-import CommentInput from "@/features/comment/components/comment-input";
-import CommentList from "@/features/comment/components/comment-list";
+import { getComments } from "@/features/comment/apis/mock";
+import { CommentInput, CommentList } from "@/features/comment/components";
 import { useAlert } from "@/hooks/use-alert";
 import { useDialog } from "@/hooks/use-dialog";
 import { useModal } from "@/hooks/use-modal";
