+++ src/features/comment/components/index.ts
@@ -1,0 +1,2 @@
+export { default as CommentInput } from "./comment-input";
+export { default as CommentList } from "./comment-list";
